i think it almost done

+++ src/pages/pengelolaan/nilai/kerja_praktek/HomeNilaiKP.jsx
@@ -3,24 +3,26 @@ import { Sidebar } from "../../../../components/sidebar/Sidebar";
 import { InfinitySpin } from "react-loader-spinner";
 import { Link, useNavigate } from "react-router-dom";
 import { useAuthState } from "react-firebase-hooks/auth";
-import { auth, db } from "../../../../utils/firebase";
+import { auth, db, storage } from "../../../../utils/firebase";
 import {
   collection,
   deleteDoc,
   doc,
   getDoc,
   onSnapshot,
+  orderBy,
   query,
   where,
 } from "firebase/firestore";
 import Swal from "sweetalert2";
+import { deleteObject, ref } from "firebase/storage";
 
-export const HomePengajuanSkripsi = () => {
+export const HomeNilaiKP = () => {
   const itemsPerPage = 5;
   const [data, setData] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
   const [isLoading, setIsLoading] = useState(true);
-  const [searchText, setSearchText] = useState("");
+  // const [searchText, setSearchText] = useState("");
   const navigate = useNavigate();
   const [user, loading] = useAuthState(auth);
 
@@ -33,55 +35,46 @@ export const HomePengajuanSkripsi = () => {
     }
     return null;
   };
+  const getPengajuanInfo = async (uid) => {
+    const userDocRef = doc(db, "pengajuan", uid);
+    const userDocSnapshot = await getDoc(userDocRef);
+
+    if (userDocSnapshot.exists()) {
+      return userDocSnapshot.data();
+    }
+    return null;
+  };
 
   useEffect(() => {
     const unsubscribe = onSnapshot(
       query(
-        collection(db, "pengajuan"),
-        where("jenisPengajuan", "==", "Skripsi")
+        collection(db, "sidang"),
+        where("jenisSidang", "==", "Kerja Praktek"),
+        orderBy("status", "asc")
       ),
       async (snapshot) => {
         const fetchedData = [];
         for (const doc of snapshot.docs) {
           const data = doc.data();
           const userInfo = await getUserInfo(data.user_uid);
-          let dosenPembimbingInfo = null;
-
-          // Cek jika pembimbing_uid tidak sama dengan "-"
-          if (data.pembimbing_uid !== "-") {
-            dosenPembimbingInfo = await getUserInfo(data.pembimbing_uid);
-          }
+          const pengajuanInfo = await getPengajuanInfo(data.pengajuan_uid);
+          const dosenPembimbingInfo = await getUserInfo(
+            pengajuanInfo.pembimbing_uid
+          );
+          const pengujiSatuInfo = await getUserInfo(data.pengujiSatu_uid);
+          const pengujiDuaInfo = await getUserInfo(data.pengujiDua_uid);
           fetchedData.push({
             id: doc.id,
             ...data,
             userInfo: userInfo,
             dosenPembimbingInfo: dosenPembimbingInfo,
+            pengujiSatuInfo: pengujiSatuInfo,
+            pengujiDuaInfo: pengujiDuaInfo,
+            pengajuanInfo: pengajuanInfo,
           });
         }
-
-        const filteredData = fetchedData.filter(
-          (item) =>
-            item.userInfo.nama
-              .toLowerCase()
-              .includes(searchText.toLowerCase()) ||
-            item.userInfo.jurusan
-              .toLowerCase()
-              .includes(searchText.toLowerCase()) ||
-            item.userInfo.nim
-              .toLowerCase()
-              .includes(searchText.toLowerCase()) ||
-            item.status.toLowerCase().includes(searchText.toLowerCase()) ||
-            item.dosenPembimbingInfo.nama
-              .toLowerCase()
-              .includes(searchText.toLowerCase()) ||
-            new Date(item.createdAt.seconds * 1000)
-              .toLocaleDateString("en-US")
-              .includes(searchText) ||
-            item.topikPenelitian
-              .toLowerCase()
-              .includes(searchText.toLowerCase())
-        );
-        setData(filteredData);
+        setData(fetchedData);
+        console.log(data);
         setIsLoading(false);
       }
     );
@@ -92,6 +85,14 @@ export const HomePengajuanSkripsi = () => {
     // Cleanup: unsubscribe when the component unmounts or when the effect re-runs
     return () => unsubscribe();
   }, [user, loading]);
+
+  const truncateTitle = (title, words = 3) => {
+    const wordsArray = title.split(" ");
+    if (wordsArray.length > words) {
+      return wordsArray.slice(0, words).join(" ") + "...";
+    }
+    return title;
+  };
 
   const handleDelete = async (id) => {
     try {
@@ -107,8 +108,24 @@ export const HomePengajuanSkripsi = () => {
       });
 
       if (result.isConfirmed) {
-        const docRef = doc(db, "pengajuan", id);
-        await deleteDoc(docRef);
+        const docRef = doc(db, "sidang", id);
+        const docSnapshot = await getDoc(docRef);
+        if (docSnapshot.exists()) {
+          const data = docSnapshot.data();
+          const persetujuanKPFileName = `persyaratan/sidangKP/formPersetujuanKP/${data.uid}`;
+          const penilaianPerusahaanFileName = `persyaratan/sidangKP/penilaianPerusahaan/${data.uid}`;
+          const pendaftaranKpFileName = `persyaratan/sidangKP/formPendaftaranKP/${data.uid}`;
+          const bimbinganKPFileName = `persyaratan/sidangKP/formBimbinganKP/${data.uid}`;
+          const sertifikatSeminarFileName = `persyaratan/sidangKP/sertifikatSeminar/${data.uid}`;
+          const sertifikatPSPTFileName = `persyaratan/sidangKP/sertifikatPSPT/${data.uid}`;
+          await deleteObject(ref(storage, persetujuanKPFileName));
+          await deleteObject(ref(storage, penilaianPerusahaanFileName));
+          await deleteObject(ref(storage, pendaftaranKpFileName));
+          await deleteObject(ref(storage, bimbinganKPFileName));
+          await deleteObject(ref(storage, sertifikatSeminarFileName));
+          await deleteObject(ref(storage, sertifikatPSPTFileName));
+          await deleteDoc(docRef);
+        }
         Swal.fire("Success", "Data Berhasil dihapus!", "success");
       }
     } catch (error) {
@@ -128,15 +145,13 @@ export const HomePengajuanSkripsi = () => {
             <Sidebar />
             <div className="flex flex-col w-full pl-[300px] overflow-y-auto pr-4 pb-4">
               <h1 className="text-2xl text-white text-center shadow-md font-bold rounded-lg p-4 m-4 mb-10 bg-slate-600">
-                Data Pengajuan Skripsi
+                Data Nilai Kerja Praktek
               </h1>
               <div className="flex items-center mt-16 mb-2 mx-2 justify-end mr-4">
                 <input
                   type="text"
                   className="px-4 py-2 border w-[400px] rounded-md drop-shadow-sm"
                   placeholder="Search..."
-                  value={searchText}
-                  onChange={(e) => setSearchText(e.target.value)}
                 />
               </div>
 
@@ -146,14 +161,16 @@ export const HomePengajuanSkripsi = () => {
                   <thead className=" shadow-sm font-extralight text-sm">
                     <tr className="">
                       <th className="p-2 px-6">No</th>
-                      <th className="p-2 px-6">Tanggal Daftar</th>
                       <th className="p-2 px-6">NIM</th>
                       <th className="p-2 px-6">Nama</th>
                       <th className="p-2 px-6">Jurusan</th>
-                      <th className="p-2 px-6">Topik Penelitian</th>
-                      <th className="p-2 px-6">Status</th>
+                      <th className="p-2 px-6">Judul</th>
                       <th className="p-2 px-6">Catatan</th>
                       <th className="p-2 px-6">Pembimbing</th>
+                      <th className="p-2 px-6 whitespace-nowrap">Penguji 1</th>
+                      <th className="p-2 px-6 whitespace-nowrap">Penguji 2</th>
+                      <th className="p-2 px-6 whitespace-nowrap">Nilai</th>
+                      <th className="p-2 px-6 whitespace-nowrap">Indeks</th>
                       <th className="p-2 px-6">Action</th>
                     </tr>
                   </thead>
@@ -165,18 +182,6 @@ export const HomePengajuanSkripsi = () => {
                       >
                         <td className="text-center">{index + 1}</td>
                         <td className="text-center">
-                          {item.createdAt &&
-                            new Date(
-                              item.createdAt.seconds * 1000
-                            ).toLocaleDateString("id-ID", {
-                              day: "2-digit",
-                              month: "2-digit",
-                              year: "numeric",
-                            })}
-
-                          {item.tanggalDaftar}
-                        </td>
-                        <td className="text-center">
                           {item.userInfo && item.userInfo.nim}
                         </td>
                         <td className="text-center whitespace-nowrap">
@@ -186,22 +191,35 @@ export const HomePengajuanSkripsi = () => {
                           {item.userInfo && item.userInfo.jurusan}
                         </td>
                         <td className="text-center p-4 whitespace-nowrap">
-                          {item.topikPenelitian}
+                          {truncateTitle(item.pengajuanInfo.judul, 3)}
                         </td>
-                        <td className="text-center">{item.status}</td>
                         <td className="text-center p-4">{item.catatan}</td>
-                        <td className="text-center p-6">
-                          {item.dosenPembimbingInfo
+                        <td className="text-center p-6 whitespace-nowrap">
+                          {item.pengajuanInfo
                             ? item.dosenPembimbingInfo.nama
                             : "-"}
+                        </td>
+                        <td className="text-center p-6 whitespace-nowrap">
+                          {item.pengujiSatuInfo
+                            ? item.pengujiSatuInfo.nama
+                            : "-"}
+                        </td>
+                        <td className="text-center p-6 whitespace-nowrap">
+                          {item.pengujiDuaInfo ? item.pengujiDuaInfo.nama : "-"}
+                        </td>
+                        <td className="text-center p-6 whitespace-nowrap">
+                          {item.nilaiAkhir ? item.nilaiAkhir : "-"}
+                        </td>
+                        <td className="text-center p-6 whitespace-nowrap">
+                          {item.indeks ? item.indeks : "-"}
                         </td>
                         <td className="text-center p-4">
                           <div className="flex">
                             <Link
-                              to={`/pengajuan-skripsi/detail/${item.id}`}
+                              to={`/kelola-nilai/kp/${item.id}`}
                               className="p-2 bg-slate-200 rounded-md hover:bg-slate-300 mr-1"
                             >
-                              Detail
+                              Nilai
                             </Link>
                             <button
                               className="p-2 bg-red-200 rounded-md hover:bg-red-300"
